requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3; python_version < "3.11"
attrs==24.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.19.1
dill==0.3.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.2.2; python_version < "3.11"
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3; (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64") and python_version < "3.13"
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.13
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.0
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.99
language-data==1.2.0
loguru==0.7.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-cupti-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-nvrtc-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-runtime-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cudnn-cu12==8.9.2.26; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufft-cu12==11.0.2.54; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-curand-cu12==10.3.2.106; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusolver-cu12==11.4.5.107; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparse-cu12==12.1.0.106; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nccl-cu12==2.18.1; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvjitlink-cu12==12.6.20; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvtx-cu12==12.1.105; platform_system == "Linux" and platform_machine == "x86_64"
openai==1.40.8
openpyxl==3.1.2
orjson==3.10.7
packaging==23.2
pandas==2.2.2
pillow==10.4.0
preshed==3.0.9
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pyyaml==6.0.2
regex==2024.7.24
requests==2.31.0
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==3.0.1
setuptools==72.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.32
srsly==2.4.8
sympy==1.13.2
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transformers==4.37.2
triton==2.1.0; platform_system == "Linux" and platform_machine == "x86_64"
typer==0.12.3
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
win32-setctime==1.1.0; sys_platform == "win32"
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
